Resource loading on startup, and a sound system
Added a global resources object that holds all sprites and sounds.
Sprites and sounds are now loaded, and their variable names declared, in
the top of spyrogame.js.
The gameStep function has been changed to wait with running it's main
code untill all sprites and sounds are loaded.
The loadSprite() function has been moved to resources.js, renamed to
addSprite(), and is now a property/method of the resources object, along
with an equivalent addSound() function.

Sounds are played with audio.playSound().
'audio' is a new global object that contains a list of sounds that are
currently playing.
A sound object returned from audio.playSound() can manipulate it's
playing sound with a set of functions.

The spyro object makes a sound when jumping, and the enemy objects makes
a sound and changes Spyro's sprite when coming near.

I tried to clean up some of the older code to make it more readable.