Polygons and Level Editor
I added polygons, and a level editor based mostly around making polygon
structures.

When you run the game, it will now ask if you want to open level editor,
or to play a level.

Two new types of resources for the polygons, textures and details, both
are images. Textures should preferably be seamless, and they don't have
any additional properties, while details are a kind of unanimated
sprites with an origin, used to spruce up the look of polygons.
I made some temporary textures/details to test the editor while making
it, and to create a couple of example "levels".

The level editor has two textfields, one to which it outputs the level
data when you tell it to save a level, and the other where you can paste
a level data string and load it into the editor.
I decided to put the example level strings inside the spyrogame.js for
now.

Spyro has a new, relatively long collision code to collide with the
polygon terrain. It should probably be turned into a general purpose
collision script that other objects can use too, but first a system that
splits the collideable lines into a grid, for otimized collision
checking, needs to be made.

To make adding new objects also to the editor more easy, I made an
objectType list, to which object names and object's constructors are
added, and to which you can assign properties that will be editable in
the level editor.

I made a polygonlevel object, which loads a level string and holds all
the polygons and objects in the level, and makes them run their step()
and draw() functions.

The level editor GUI is html, ugly/messy, and currently in the main
.html file because I couldn't think of how to easily include it
elsewise.
The level editor code is at around 1300 lines, and a bit messy, but I
tried to clean some of it, and to add comments. The editor works,
though, and it fulfills it's purpose.

I might have forgotten some important points, since there were a lot of
changes and new things.