Collision grid
I made the polygonLevel object store the indexes of it's polygon's lines
in a grid.
The polygons also calculate their individual bounding boxes.
Each cell in the grid contains a list of objects with :
-the index polygon a line belongs to
-the index of the line's first endpoint in that polygon's list of points

Also:
A function to extract a part of the level, a list like the one mentioned
above, but containing only lines that touch a specified circle.
A duplicate of objectXlevelCollision() but which tests only with such a
list.
A simple circleXlevelPart collision test.

Changed some of Spyro's code to use the grid system, made him slow
down/be able to jump only when he has ground under his feet. Made him do
some debug drawing, mainly just to illustrate part of the grid system.

Edited the enemies a bit, made them bump around after Spyro.