Optimized polygon drawing
The polygons now cache their appearance to their own canvases, which are
drawn to the main canvas. This should result in much better performance.

Changed the big level a bit to work better with the polygon caching.
( I split the huge polygon into smaller ones).

Removed some unecessary debug drawing from my previous commits.

Made a small change to timer.js to make it reduce it's delay faster
because when I ran the game in FF, and changed to another tab, it
thought the game was lagging, thus added a lot of delay. It took ages
for it to go back to normal, so I made it a little quicker. Hope it
doesn't cause problems with actual lag issues.