Animated Spyro and gliding, falling, etc...
Spyro now has some of his sprites, and a more complex behavior including
gliding, hovering, falling, running and breaking/slowing down, he
rotates with the terrain, and he no longer moves off the ground when
running down slopes.

For this - the rotation and sticking to the ground in particular - I had
to change the onGround detection to use a downward ray intersection test
instead if the circleXterrain test previously used.