(Incomplete) Chest collision, backgrounds & portal
I added a chest collision function, separate from the main
collision/reaction function, as I don't think all objects need to take
chests into consideration. There's still an issue when jumping up
through platforms and into chests.

Added a background image, a portal(only visual at the moment), and a
badly recolorized version of the background image to be displayed
through the portal.
I also started preparing for portals actually working by making levels
have an id( a.k.a. it's index in the list of level data), using this id
to draw a level's correct background, and portals know the id of the
levels they're leading to.