Charging!
Implemented charging. This made spyro.js a bit more confusing, partially
because the charge state "overlaps" with a couple of other states:
jumping and "no state"(on ground, walking).

Changed the controls scheme a bit:
Left and Right are still sideways movement, but S is now jump, while A
is charge and D is flame. I think W can work as Triangle, for looking
around and talking. (changed to this in moneybags.js)
Made a few named key clones, like jumpKey = sKey.

Also made gems unpickable the first few steps they exist, so that they
get time to pop up out of enemies and chests.